feat: implement authentication middleware and secure API routes; add login and logout functionality; enhance UI components for user interaction

+++ app/api/drinks/route.ts
@@ -1,12 +1,21 @@
-import { NextResponse } from 'next/server';
+import { NextResponse, NextRequest } from 'next/server';
 import { prisma } from '@/lib/prisma';
+import { isAuthenticated } from '@/app/auth/isAuthenticated';
 
-export async function GET() {
+export async function GET(req: NextRequest) {
+  if (!isAuthenticated(req)) {
+      return NextResponse.json({ error: 'Unauthorized' }, { status: 401 });
+    }
+
   const drinks = await prisma.drink.findMany();
   return NextResponse.json(drinks);
 }
 
-export async function POST(req: Request) {
+export async function POST(req: NextRequest) {
+  if (!isAuthenticated(req)) {
+    return NextResponse.json({ error: 'Unauthorized' }, { status: 401 });
+  }
+
   const body = await req.json();
   const drink = await prisma.drink.create({
     data: {
@@ -18,7 +27,11 @@ export async function POST(req: Request) {
   return NextResponse.json(drink, { status: 201 });
 }
 
-export async function DELETE(req: Request) {
+export async function DELETE(req: NextRequest) {
+  if (!isAuthenticated(req)) {
+    return NextResponse.json({ error: 'Unauthorized' }, { status: 401 });
+  }
+
   const body = await req.json();
   const drink = await prisma.drink.delete({
     where: {
